+++ Runtime/Localized Reference/LocalizedString.cs
@@ -52,6 +52,9 @@ namespace UnityEngine.Localization
         /// <param name="handler"></param>
         public void RegisterChangeHandler(ChangeHandler handler)
         {
+            if (handler == null)
+                throw new ArgumentNullException(nameof(handler));
+
             LocalizationSettings.ValidateSettingsExist();
             m_ChangeHandler = handler ?? throw new ArgumentNullException(nameof(handler), "Handler must not be null");
             LocalizationSettings.SelectedLocaleChanged += HandleLocaleChange;
@@ -67,6 +70,7 @@ namespace UnityEngine.Localization
             LocalizationSettings.ValidateSettingsExist();
             LocalizationSettings.SelectedLocaleChanged -= HandleLocaleChange;
             m_ChangeHandler = null;
+            ClearLoadingOperation();
         }
         
         /// <summary>
@@ -135,10 +139,8 @@ namespace UnityEngine.Localization
         void HandleLocaleChange(Locale _)
         {
             // Cancel any previous loading operations.
-            if (CurrentLoadingOperation != null && !CurrentLoadingOperation.Value.IsDone)
-            {
-                CurrentLoadingOperation.Value.Completed -= AutomaticLoadingCompleted;
-            }
+            ClearLoadingOperation();
+
             m_Entry = null;
             m_Table = null;
 
@@ -160,6 +162,15 @@ namespace UnityEngine.Localization
             m_Entry = loadOperation.Result.Entry;
             m_Table = loadOperation.Result.Table;
             RefreshString();
+        }
+
+        void ClearLoadingOperation()
+        {
+            if (m_CurrentLoadingOperation.HasValue)
+            {
+                m_CurrentLoadingOperation.Value.Completed -= AutomaticLoadingCompleted;
+                m_CurrentLoadingOperation = null;
+            }
         }
     }
 }