+++ Runtime/Operations/PreloadDatabaseOperation.cs
@@ -9,7 +9,7 @@ namespace UnityEngine.Localization
 {
     class PreloadDatabaseOperation<TTable, TEntry> : AsyncOperationBase<LocalizedDatabase<TTable, TEntry>>
         where TTable : LocalizedTableT<TEntry>
-        where TEntry : TableEntry, new()
+        where TEntry : TableEntry
     {
         LocalizedDatabase<TTable, TEntry> m_Database;
         AsyncOperationHandle<IList<IResourceLocation>> m_LoadResourcesOperation;